+++ CHANGELOG.md
@@ -5,7 +5,12 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.2 - Unreleased
+## Version 0.2.3 - Unreleased
+
+### Fixed
+* Issue in SensorChan spec where duplicate sensors were present in concise codes
+
+## Version 0.2.2 - Released 2022-09-28
 
 ### Added
 * Added `__json__` to sensorchan specs